simulators/my_simulator/pyproject.toml
[tool.poetry]
name = "my_simulator"
version = "0.1.0"
description = "A Composabl gRPC wrapper around a dummy Simulator"
authors = ["Octavio Santiago <octaviobomfim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.49.1"
grpcio-tools = "~1.49.1"
grpcio-reflection = "~1.49.1"
opencv-python = "^4.7.0.72"
gymnasium = {version = "~0.26.3"}
#composabl = {path = "../../composabl/sdks/python_sdk", develop = true}
asyncio = "^3.4.3"
scipy = {version = "^1.10.1", python = ">=3.8,<3.12" }
composabl-dev = "^0.2.3.dev30"

[tool.poetry.scripts]
main = "my_simulator.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
